Add option to purge apps and purge apps with --all
One can now choose to purge apps with 'scitran purge --apps'. This
addresses the following issue:

When purging an instance the apps were not being purged, which was
causing apps to not be bootstrapped when starting the instance from
"scratch". This, in turn, was causing jobs not to be started as there
were no apps in the DB.

This fixes that issue by removing/purging the apps directory, which
triggers a bootstrap prompt for apps when starting the app.

In this commit are also a few edits to the user prompts.